Moved play function to funcs folder
